Rename ut_confidence.py to confidence.py
